refactor: remove unused imports
