+++ src/App.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
-import { RouterLink, RouterView } from 'vue-router'
-import LandingView from './views/LandingView.vue';
+import { RouterView } from 'vue-router'
 </script>
 
 <template>
