+++ src/views/WeatherView.vue
@@ -1,10 +1,8 @@
 <script setup lang="ts">
-import { ref } from 'vue';
-import { RouterLink } from 'vue-router';
 import { storeToRefs } from 'pinia';
 import { useWeatherStore } from '@/stores/weather';
 import parseWeatherCode from '../utils/weatherCodes';
-import Searchbar from '@/components/Searchbar.vue';
+import Searchbar from '@/components/LocationSearchbar.vue';
 
 const weatherStore = useWeatherStore();
